# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Loan App API",
        "description": "A simple loan management API built with Laravel.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "description": "Register a new user and issue an API token.",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ada Lovelace"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ada@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123!"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered, token issued"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log in",
                "description": "Authenticate a user and issue an API token.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ada@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authenticated, token issued"
                    },
                    "422": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get the authenticated user",
                "description": "Return the authenticated user.",
                "operationId": "9d9bc556f927575aa492a88c28fe5908",
                "responses": {
                    "200": {
                        "description": "The authenticated user"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log out (revoke current token)",
                "description": "Revoke the current access token (log out).",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "204": {
                        "description": "Token revoked"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/loans": {
            "get": {
                "tags": [
                    "Loans"
                ],
                "summary": "Retrieve all loans",
                "description": "Retrieve all loans (most recent first, paginated).",
                "operationId": "53fdef403307586122900dc3b1d07368",
                "responses": {
                    "200": {
                        "description": "Paginated list of loans",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Loan"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Loans"
                ],
                "summary": "Create a loan",
                "description": "Create a loan.",
                "operationId": "92bd5c653a9a8345b4f161d5ba3bbd71",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoanInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Loan created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Loan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/loans/me": {
            "get": {
                "tags": [
                    "Loans"
                ],
                "summary": "Retrieve the authenticated user's loans",
                "description": "Retrieve all loans belonging to the authenticated user.",
                "operationId": "c7b067cb44bf92dbf66a8b5c4cbcb367",
                "responses": {
                    "200": {
                        "description": "Paginated list of the current user's loans",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Loan"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/loans/{loan}": {
            "get": {
                "tags": [
                    "Loans"
                ],
                "summary": "Retrieve a single loan",
                "description": "Retrieve a single loan.",
                "operationId": "39a4e10f305b5be4c3e4eba31a94777a",
                "parameters": [
                    {
                        "name": "loan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The loan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Loan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Loan not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Loans"
                ],
                "summary": "Update a loan",
                "description": "Update a loan.",
                "operationId": "d1a9dfd54e508debb4769bc440e372f8",
                "parameters": [
                    {
                        "name": "loan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoanInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Loan updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Loan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Loan not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Loans"
                ],
                "summary": "Delete a loan",
                "description": "Delete a loan.",
                "operationId": "6d80bd9e85d71ce0f714269e753f85fb",
                "parameters": [
                    {
                        "name": "loan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Loan deleted"
                    },
                    "404": {
                        "description": "Loan not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Loan": {
                "title": "Loan",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "borrower_name": {
                        "type": "string",
                        "example": "Ada Lovelace"
                    },
                    "borrower_email": {
                        "type": "string",
                        "format": "email",
                        "example": "ada@example.com"
                    },
                    "amount": {
                        "type": "string",
                        "example": "5000.00"
                    },
                    "interest_rate": {
                        "type": "string",
                        "example": "12.50"
                    },
                    "term_months": {
                        "type": "integer",
                        "example": 24
                    },
                    "status": {
                        "type": "string",
                        "example": "pending",
                        "enum": [
                            "pending",
                            "approved",
                            "active",
                            "paid",
                            "rejected"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "LoanInput": {
                "title": "Loan input",
                "required": [
                    "borrower_name",
                    "borrower_email",
                    "amount",
                    "interest_rate",
                    "term_months"
                ],
                "properties": {
                    "borrower_name": {
                        "type": "string",
                        "example": "Ada Lovelace"
                    },
                    "borrower_email": {
                        "type": "string",
                        "format": "email",
                        "example": "ada@example.com"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 5000
                    },
                    "interest_rate": {
                        "type": "number",
                        "format": "float",
                        "example": 12.5
                    },
                    "term_months": {
                        "type": "integer",
                        "example": 24
                    },
                    "status": {
                        "type": "string",
                        "example": "pending",
                        "enum": [
                            "pending",
                            "approved",
                            "active",
                            "paid",
                            "rejected"
                        ]
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum personal access token. Send as: Authorization: Bearer {token}",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Loans",
            "description": "Loans"
        }
    ]
}